api/requirements.txt
# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP requests and HTML parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0

# Airtable integration
pyairtable>=2.1.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Rate limiting
slowapi>=0.1.9

# Logging and utilities
python-json-logger>=2.0.7

# Celery and message queue
celery[redis]>=5.3.0
redis>=5.0.0
flower>=2.0.0
kombu>=5.3.0

# RabbitMQ client for WhatsApp service
pika>=1.3.0

# PostgreSQL database
psycopg2-binary>=2.9.9

# Authentication and security
PyJWT>=2.8.0
bcrypt>=4.1.2
python-slugify>=8.0.1
email-validator>=2.0.0

